# Optimize ConfigurationPropertyName