@wilkinsona 2019-01-25
#### Main method

| | Baseline | New |
| - | - | - |
| 1. | 1 | 1 |
| 1. | 1 | 1 |
| 1. | 1 | 1 |
| 1. | 1 | 1 |
| 1. | 1 | 1 |