@wilkinsona 2019-01-25
Wow! This makes a significant difference to the time it takes to bind `InitializrProperties` with the large YAML file. I used the following app:

```java
@EnableConfigurationProperties(InitializrProperties.class)
public class BinderPerformanceApplication {

	public static void main(String[] args) {
		SpringApplication.run(BinderPerformanceApplication.class, args);
	}

}
```

The times in the following table are the JVM running time that's logged on startup:

| | Baseline | New |
| - | -: | -: |
| | 2.279 | 1.54 |
| | 2.177 | 1.42 |
| | 2.151 | 1.555 |
| | 2.703 | 1.459 |
| | 2.445 | 1.452 |
| | 2.358 | 1.613 |
| | 2.238 | 1.745 |
| | 2.445 | 1.763 |
| | 2.296 | 1.422 |
| | 2.486 | 1.456 |
| Mean | 2.358 | 1.543 |
| Range | 0.552 | 0.343 |

It's 800ms faster, reducing startup time by 34%.