@dreis2211 2019-01-25
Wow. that's even more than on my machine. I'm glad that this seems to be a worthwhile improvement so far.