@wilkinsona 2019-01-25
And here are equivalent numbers for the same app with a much smaller YAML file:

|       | Baseline |  New  |
| ----- | -------: | ----: |
|       |    0.780 | 0.744 |
|       |    0.742 | 0.729 |
|       |    0.759 | 0.727 |
|       |    0.756 | 0.735 |
|       |    0.767 | 0.748 |
|       |    0.747 | 0.747 |
|       |    0.757 | 0.738 |
|       |    0.755 | 0.730 |
|       |    0.770 | 0.735 |
|       |    0.767 | 0.740 |
|  Mean |    0.760 | 0.737 |
| Range |    0.038 | 0.021 |

New's faster, but by nowhere near as much. That's to be expected with far fewer properties to compare.

This change will be slower in scenarios where there are a lot of properties with the same number of elements that differ towards the beginning rather than at the end. I think configuration of that kind will be the exception rather than the rule, and even when the configuration has that structure, the performance will be no worse than it is today for what I believe is the more common case.