@dreis2211 2019-01-25
@wilkinsona Thanks for checking the PR btw and running further benchmarks.