@wilkinsona 2019-01-28
Thanks very much for the PR, @dreis2211. This one is now in master and is responsible for the lion's share of an [almost 40% reduction in startup time](
https://github.com/spring-projects/spring-boot/issues/15760#issuecomment-458102857).